Populate the paths where the plugins live when the module is instanciated.
I had a problem using wmiirc-lua when I installed it on a machine which I don't
manage myself.  It ended up that in the wmii.lua file the plugin_path variable
is generated with hard-coded directories which assume installation to "/usr/*"
or "/usr/local".  This patch just generates the plugin paths based on the
package.path variable.

As a side note, it also stores each path in a table rather than as one long
string.  This way they don't have to be re-split everytime a plugin is loaded.